requirements.txt
asgiref             3.7.2
bakery              1.0.0
beautifulsoup4      4.12.2
colorama            0.4.6
distlib             0.3.7
dj-database-url     2.1.0
Django              4.2.6
django-bootstrap-v5 1.0.11
django-filter       23.3
django-pytest       0.2.0
djangorestframework 3.14.0
filelock            3.12.3
gunicorn            21.2.0
iniconfig           2.0.0
model-bakery        1.15.0
packaging           23.2
pip                 23.3.1
platformdirs        3.10.0
pluggy              1.3.0
pytest              7.4.2
pytest-django       4.5.2
python-decouple     3.8
pytz                2023.3.post1
setuptools          65.5.0
soupsieve           2.5
sqlparse            0.4.4
typing_extensions   4.8.0
tzdata              2023.3
virtualenv          20.24.5
whitenoise          6.6.0
PS C:\Users\Notebook\Documents\humberto\projeto_final> pip freeze
asgiref==3.7.2
bakery==1.0.0
beautifulsoup4==4.12.2
colorama==0.4.6
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.6
django-bootstrap-v5==1.0.11
django-filter==23.3
django-pytest==0.2.0
djangorestframework==3.14.0
filelock==3.12.3
gunicorn==21.2.0
iniconfig==2.0.0
model-bakery==1.15.0
packaging==23.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-django==4.5.2
python-decouple==3.8
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.5
whitenoise==6.6.0